@@ -12,18 +12,18 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.0",
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/forms": "^5.2.0",
-    "@angular/http": "^5.2.0",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/router": "^5.2.0",
-    "@ng-idle/core": "^2.0.0-beta.15",
-    "@ng-idle/keepalive": "^2.0.0-beta.15",
-    "@ngrx/store": "^5.2.0",
+    "@angular/animations": "7.1.3",
+    "@angular/common": "7.1.3",
+    "@angular/compiler": "7.1.3",
+    "@angular/core": "7.1.3",
+    "@angular/forms": "7.1.3",
+    "@angular/http": "7.1.3",
+    "@angular/platform-browser": "7.1.3",
+    "@angular/platform-browser-dynamic": "7.1.3",
+    "@angular/router": "7.1.3",
+    "@ng-idle/core": "^6.0.0-beta.3",
+    "@ng-idle/keepalive": "^6.0.0-beta.3",
+    "@ngrx/store": "^6.1.2",
     "angular2-chartjs": "^0.5.1",
     "angular2-draggable": "^1.4.2",
     "angular2-moment": "^1.9.0",
@@ -38,13 +38,15 @@
     "ngx-cookie-service": "^1.0.10",
     "ngx-treeview": "^6.0.1",
     "popper.js": "^1.14.3",
-    "rxjs": "^5.5.6",
-    "zone.js": "^0.8.19"
+    "rxjs": "^6.3.3",
+    "tslib": "^1.9.0",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "~1.7.4",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
+    "@angular-devkit/build-angular": "~0.11.0",
+    "@angular/cli": "^7.1.3",
+    "@angular/compiler-cli": "7.1.3",
+    "@angular/language-service": "7.1.3",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
     "@types/jspdf": "^1.1.31",
@@ -57,9 +59,9 @@
     "karma-coverage-istanbul-reporter": "^1.2.1",
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
+    "protractor": "^5.4.1",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "~2.5.3"
+    "typescript": "3.1.6"
   }
 }
