@@ -1,9 +1,11 @@
 import {Injectable} from '@angular/core';
 import {API_URLS} from '../config/api.url.config';
 import {HttpClient} from '@angular/common/http';
-import {Observable} from 'rxjs/Observable';
+import {Observable} from 'rxjs';
 
-@Injectable()
+@Injectable({
+  providedIn: 'root',
+})
 export class MovementService {
 
   constructor(private http: HttpClient){}
