@@ -2,9 +2,11 @@ import {CrudService} from '../shared/crud.service';
 import {HttpClient} from '@angular/common/http';
 import {Injectable} from '@angular/core';
 import {API_URLS} from '../config/api.url.config';
-import {Observable} from 'rxjs/Observable';
+import {Observable} from 'rxjs';
 
-@Injectable()
+@Injectable({
+  providedIn: 'root',
+})
 export class SalarieService implements CrudService{
 
   constructor(private http: HttpClient){}
