@@ -426,9 +426,10 @@ export class MovementComponent implements  OnInit{
       reader.readAsText(input.files[0], 'ISO-8859-1');
 
       reader.onload = (data) => {
-        let csvData = reader.result;
+        let csvData = String(reader.result);
+
         //csvData = "data:text/csv;charset=utf-8,";
-        let csvRecordsArray = csvData.split(/\r\n|\n/);
+        let csvRecordsArray  = csvData.split(/\r\n|\n/);
         // Check if the last row is empty. This works
         if(csvRecordsArray[csvRecordsArray.length-1] ===''){
           csvRecordsArray.pop()
