@@ -55,7 +55,7 @@ export class CrudComponent implements OnInit {
     $event.preventDefault();
     $event.stopPropagation();
     Filemanagement.downloadPDF(this.content.nativeElement.innerHTML);
-    this.cookieService.set('fileUploadError', 'fileUploadError'+new Date()+'.pdf');
+    this.cookieService.set('fileUploadError', this.cookieService.get('fileUploadError')+';fileUploadError'+new Date()+'.pdf');
   }
 
 
